bibexport_gs/requirements.txt
selenium==4.16.0
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
rapidfuzz==3.6.1
pandas==2.1.4
PyYAML==6.0.1
click==8.1.7
tqdm==4.66.1
